% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpull.R
\name{xpull}
\alias{xpull}
\title{xpull}
\usage{
xpull(table, auth, use_ext = FALSE, format = "csv", mart = "NCOV")
}
\arguments{
\item{table}{- a string representing the name of the table in xmart}

\item{auth}{- authorisation package}

\item{use_ext}{- use cached data? (downloads faster, slightly out of date)}

\item{format}{- the format of the table to be pulled - can take json or csv (csv)}

\item{mart}{- which mart should be used? NCOV by default}
}
\description{
xpull
}
\details{
code to pull data from XMART into R via csv - FASTER, doesn't give column types
}
\author{
HLS
}
