% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpull_csv.R
\name{xpull_csv}
\alias{xpull_csv}
\title{xpull_csv}
\usage{
xpull_csv(table, auth, use_ext = FALSE, mart = "NCOV")
}
\arguments{
\item{table}{- the name of the table on xmart to be used}

\item{auth}{- authorisation package}

\item{mart}{- which mart should be used? NCOV by default}

\item{cached}{- use cached data? (downloads faster, slightly out of date)}
}
\description{
xpull_csv
}
\details{
code to pull data from XMART into R via csv - FASTER, doesn't give column types
}
\author{
HLS
}
