% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpull_json.R
\name{xpull_json}
\alias{xpull_json}
\title{xpull_json}
\usage{
xpull_json(table, auth, use_ext = FALSE, mart = "NCOV")
}
\arguments{
\item{table}{- what table should be pulled}

\item{auth}{- authorisation package}

\item{use_ext}{- use cached data? (downloads faster, slightly out of date)}

\item{mart}{- which mart should be used? NCOV by default}
}
\description{
xpull_json
}
\details{
code to pull data from XMART into R - NOT EXPORTED
}
\author{
HLS
}
